abstracted logic from schemaDisplay to createNewNodesAndEdges
